refactor(middleware): split audit/client-ip and support CIDR allowlist
